requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
statsmodels
sqlalchemy
psycopg[binary]
pytest
psycopg2-binary
plotly
dash
